[CM] Fix some coverity issue
Fix memory leak issue
Fix no teminate break issue
Fix dead logic issue
Fix allocate insufficient memory issue

+++ media_driver/agnostic/common/cm/cm_surface_3d_rt.cpp
@@ -126,7 +126,7 @@ CM_RT_API int32_t CmSurface3DRT::WriteSurface( const unsigned char* sysMem,
             return CM_INVALID_ARG_VALUE;
     }
 
-    uSizeInBytes = m_width * m_height * m_depth * pixel;
+    uSizeInBytes = (uint64_t)(m_width * m_height * m_depth * pixel);
     if (sysMemSize < uSizeInBytes)
     {
         CM_ASSERTMESSAGE("Error: Invalid copy size.")
@@ -257,7 +257,7 @@ CM_RT_API int32_t CmSurface3DRT::ReadSurface( unsigned char* sysMem, CmEvent* ev
             return CM_INVALID_ARG_VALUE;
     }
 
-    uSizeInBytes = m_width * m_height * m_depth * pixel;
+    uSizeInBytes = (uint64_t)(m_width * m_height * m_depth * pixel);
     if (sysMemSize < uSizeInBytes)
     {
         CM_ASSERTMESSAGE("Error: Invalid copy size.")
